requirements.txt
cryptography==3.1.1
django==1.11.15
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
pytest==3.5.1
natsort==5.5.0
PyMySQL==0.9.3
glob2==0.6
